Some more examples of packing
I was feeling rusty so I thought I would do some more work with the example Finding Fortune had made, just for the fun of it.
So I wrapped it into a nice class and made a function that can take any number of inputs to output one packed integer.
Hope someone finds it useful or interesting.
 - frozen_dude